misc: move clients outside of the core folder and into the client folder
